Refactor animation module by introducing AnimationConfig, AnimationMode, and LoopMode enums, and enhancing Animatable trait functionality
